Set default locale to en-GB
Issue: The melius devices we support are mainly sold outside
of the US, and since en-US is the default locale, that means
time and date format are 12H and MMDDYY. This is mildly
annoying for those that prefer English as the main language,
but wants international time and date format.

Solution: Set default locale to en-GB, so the majority of users
don't need to manually change the language.

Change-Id: I43b5b4eb876b92d6f57a2c0c9e3083652cb562ee